refactor: delete LinkController and move its methods to TaskController to better abstraction (still maintain LinkService)
