+++ src/main/java/com/romm/todopp/controller/TaskController.java
@@ -1,15 +1,22 @@
 package com.romm.todopp.controller;
 
+import java.util.List;
+import java.util.Map;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.server.ResponseStatusException;
+import org.springframework.web.servlet.ModelAndView;
 
 import com.romm.todopp.entity.Link;
 import com.romm.todopp.entity.Task;
+import com.romm.todopp.entity.TaskList;
 import com.romm.todopp.service.LinkService;
+import com.romm.todopp.service.TaskListService;
 import com.romm.todopp.service.TaskService;
 
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 
@@ -19,13 +26,14 @@ import org.springframework.web.bind.annotation.PostMapping;
 @Controller @RequestMapping("/task")
 public class TaskController {
 
+    @Autowired TaskListService taskListService;
     @Autowired TaskService taskService;
     @Autowired LinkService linkService;
     
     @PostMapping("/create/{taskListId}") // refatorar isso aqui pra n precisar enviar o id dpss
     public String create(Task task, @PathVariable Long taskListId) throws ResponseStatusException {
-        taskService.create(task, taskListId);
-        linkService.create(task, taskListId);
+        var taskList = taskListService.read(taskListId);
+        taskService.create(task, taskList);
         return "redirect:/lists/" + taskListId;
     }
 
@@ -56,7 +64,27 @@ public class TaskController {
         
         return "redirect:/lists/" + taskListId;
     }
+
+    @GetMapping("/link/{taskId}/backto/{originTaskListId}")
+    public ModelAndView linkTask(@PathVariable Long taskId, @PathVariable Long originTaskListId) {
+        List<TaskList> taskLists = taskListService.findAll();
+        return new ModelAndView("link/linktask", Map.of("taskLists", taskLists, "taskId", taskId, "originTaskListId", originTaskListId));
+    }
+
+    @PostMapping("/{taskId}/linkto/{taskListId}/backto/{originTaskListId}")
+    public String linkTask(@PathVariable Long taskId, @PathVariable Long taskListId, @PathVariable Long originTaskListId) {
+        var task = taskService.read(taskId);
+        var taskList = taskListService.read(taskListId);
+        linkService.create(task, taskList);
+        
+        return "redirect:/lists/" + originTaskListId;
+    }
     
-    
+    @PostMapping("/{taskId}/unlink-of/{taskListId}")
+    public String unlink(@PathVariable Long taskId, @PathVariable Long taskListId) {
+        taskService.unlink(taskId, taskListId);
+        
+        return "redirect:/lists/" + taskListId;
+    }
     
 }
