+++ src/main/java/com/romm/todopp/service/LinkService.java
@@ -18,13 +18,10 @@ import com.romm.todopp.repository.LinkRepository;
 public class LinkService {
     
     @Autowired LinkRepository linkRepository;
-    @Autowired TaskListService taskListService;
-    @Autowired TaskService taskService;
 
-    public Link create(Task task, Long taskListId) {
-        if (alreadyExists(taskListId, task.getId())) throw new AlreadyLinkedException("A link between those already exists.", null);
+    public Link create(Task task, TaskList taskList) {
+        if (alreadyExists(taskList.getId(), task.getId())) throw new AlreadyLinkedException("A link between those already exists.", null);
 
-        TaskList taskList = taskListService.read(taskListId);
         Link link = new Link();
         link.setTask(task);
         link.setTaskList(taskList);
@@ -40,19 +37,6 @@ public class LinkService {
 
     public Link read(Long id) {
         return findOr404(id);
-    }
-
-    public void unlink(Long taskId, Long taskListId) throws ResponseStatusException {
-        Link link = findOr404(taskListId, taskId);
-        deleteAndCheckIfTaskIfOrphan(link);
-    }
-
-    public void deleteAndCheckIfTaskIfOrphan(Link link) {
-        Task task = link.getTask();
-        if (task.getLinks().size()==1) {
-            taskService.delete(task.getId());
-        }
-        delete(link);
     }
 
     public Link delete(Link link) {
